update migratrions data type float(3)

+++ misc/migrations/20240121144158_create_table_wallets.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS wallets (
     id VARCHAR PRIMARY KEY UNIQUE NOT NULL,
     customer_id VARCHAR NOT NULL REFERENCES customers(id),
-    balance FLOAT DEFAULT 0
+    balance FLOAT(3) DEFAULT 0
 );
 
 INSERT INTO wallets(id, customer_id)
